third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/coop-navigated-popup.https-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Test named 'Open a popup to a document without COOP, then navigate it to a document with' specified 1 'cleanup' function, and 1 failed.
FAIL Open a popup to a document without COOP, then navigate it to a document with assert_equals: expected 0 but got 36
Harness: the test ran to completion.
